fix(config): continue to read after overly long line
The update to Zig 0.15.2
(https://github.com/open-telemetry/opentelemetry-injector/pull/152)
broke an edge case in the config parser: After encountering a overly
long line, the config parser would stop reading the remaining lines of
the config file, and instead of only ignoring the overly long line,
would ignore the rest of the file.

While PR #152 tried to mitigate that by doubling the buffer size (and
thus the problem only would occur on lines over 16k characters), it
nevertheless only makes the bug more unlikely, but not impossible.

The better pattern to use here is to handle error.StreamTooLong inside
the while-loop, so that the while-loop is not aborted by going into its
else-branch.